navigator/backend/pyproject.toml
[project]
name = "navigator-backend"
version = "0.1.0"
description = "Navigator - Voice-first web accessibility agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "browser-use>=0.1.40",
    "browser-use-sdk>=3.3.1",
    "langchain-google-genai>=2.0.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.1.0",
    "httpx>=0.27.0",
    "playwright>=1.44.0",
    "pydantic>=2.7.0",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = []

[dependency-groups]
dev = []
